Allow date recognition by JackBean ChatBot
