+++ data/JackBeanStorage.txt
@@ -1,3 +1,3 @@
-D | my deadline | 82023e |  
-T | jasdvk |  
-E | myevent | 234 | 3524w9 | X
+T | sgasgjksdgn | X
+D | dsofajsio | 2024-08-31 |  
+E | jdosgjod | 2024-09-04 | 2024-09-05 | X
