+++ src/main/java/JackBean.java
@@ -42,79 +42,6 @@ public class JackBean {
         }
     }
 
-    // Partial solution reused from course website
-    // extra comment here to aid with tagging A-Classes
-    public static class Task {
-        protected String description;
-        protected boolean isDone;
-
-        public Task(String description) {
-            this.description = description;
-            this.isDone = false;
-        }
-
-        public String getStatusIcon() {
-            return (isDone ? "X" : " "); // mark done task with X
-        }
-
-        public void markAsDone() {
-            this.isDone = true;
-        }
-
-        public void markAsUndone() {
-            this.isDone = false;
-        }
-
-        @Override
-        public String toString() {
-            return "[" + getStatusIcon() + "] " + description;
-        }
-    }
-
-    // this class was generated by github copilot
-    public static class Deadline extends Task {
-        protected String by;
-
-        public Deadline(String description, String by) {
-            super(description);
-            this.by = by;
-        }
-
-        @Override
-        public String toString() {
-            return "[D][" + getStatusIcon() + "] " + description + " (by: " + by + ")";
-        }
-    }
-
-    // this class was generated by github copilot, but had to code from and to functionality
-    public static class Event extends Task {
-        protected String from;
-        protected String to;
-
-        public Event(String description, String from, String to) {
-            super(description);
-            this.from = from;
-            this.to = to;
-        }
-
-        @Override
-        public String toString() {
-            return "[E][" + getStatusIcon() + "] " + description + " (from: " + from + " to: " + to + ")";
-        }
-    }
-
-    // this class was generated by github copilot
-    public static class Todo extends Task {
-        public Todo(String description) {
-            super(description);
-        }
-
-        @Override
-        public String toString() {
-            return "[T][" + getStatusIcon() + "] " + description;
-        }
-    }
-
     public static void main(String[] args) {
         System.out.println(horizontalLine);
         System.out.println(greeting);
