+++ src/main/java/Todo.java
@@ -1,0 +1,10 @@
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T][" + getStatusIcon() + "] " + description;
+    }
+}
